computer_use_demo/pyproject.toml
[project]
name = "computer-use-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.39.0",
    "anyio>=4.6.2.post1",
    "bandit>=1.7.10",
    "bs4>=0.0.2",
    "ftfy>=6.3.1",
    "icecream>=2.1.3",
    "langchain-ai>=0.1.1",
    "langchain-community>=0.3.5",
    "mypy>=1.13.0",
    "openai>=1.55.0",
    "pandas>=2.2.3",
    "playwright>=1.49.0",
    "pyautogui>=0.9.54",
    "pydeps>=2.0.1",
    "pysnooper>=1.2.1",
    "python-dotenv>=1.0.1",
    "ruff>=0.7.4",
    "tenacity>=9.0.0",
    "types-beautifulsoup4>=4.12.0.20241020",
    "types-pyautogui>=0.9.3.20240406",
    "unimport>=1.2.1",
]
